Refactor tasklist into its own class
Tasklist in Marvin uses an array of strings to hold the tasklist.

To account for evolving requirements (todos, deadlines), encapsulating
behaviour of the tasklist is desirable as it grows in complexity

Create a class TodoList to hold the list and associated methods. Create
Todo class to keep Todo-related functions together. Refactor Marvin to
use the TodoList and Todo classes in its operations.