Modify Command behaviour to allow GUI usage
Command:execute executes a given command and then prints a message
out to the Ui.

With the introduction of the GUI, there needs to be a mechanism
to obtain a string-based reply from Marvin. Additionally, there
cannot be any terminal color codes encoded into the string, and
there should still be CLI support for the user.

Let's,
* Introduce a CommandResult class to encapsulate both printable
    output for the CLI and raw output for the GUI
* Modify Command and subclasses to return CommandResult
* Modify CLI run to include a call to CommandResult::printResponse
    to print out the information to the CLI
* Introduce extra methods in Personality to return uncolored
    output for the use of GUI display